StochasticBlockModelVariants

Documentation for StochasticBlockModelVariants.

StochasticBlockModelVariants.StochasticBlockModelVariants

StochasticBlockModelVariants.AbstractLatents

StochasticBlockModelVariants.AbstractObservations

StochasticBlockModelVariants.AbstractSBM

StochasticBlockModelVariants.CSBM

StochasticBlockModelVariants.GLMSBM

StochasticBlockModelVariants.LatentsCSBM

StochasticBlockModelVariants.LatentsGLMSBM

StochasticBlockModelVariants.MarginalsCSBM

StochasticBlockModelVariants.MarginalsGLMSBM

StochasticBlockModelVariants.ObservationsCSBM

StochasticBlockModelVariants.ObservationsGLMSBM

Base.length

Base.rand

StochasticBlockModelVariants.affinities

StochasticBlockModelVariants.average_degree

StochasticBlockModelVariants.communities_snr

StochasticBlockModelVariants.continuous_overlap

StochasticBlockModelVariants.discrete_overlap

StochasticBlockModelVariants.effective_snr

StochasticBlockModelVariants.evaluate_amp

StochasticBlockModelVariants.fraction_observed

StochasticBlockModelVariants.nb_features

StochasticBlockModelVariants.run_amp

StochasticBlockModelVariants.sample_graph

StochasticBlockModelVariants.sample_mask

StochasticBlockModelVariants.StochasticBlockModelVariants — Module

StochasticBlockModelVariants

A package for inference in SBMs with node features using message-passing algorithms.

Exports

AbstractSBM

CSBM

GLMSBM

GaussianWeightPrior

LatentsCSBM

LatentsGLMSBM

ObservationsCSBM

ObservationsGLMSBM

RademacherWeightPrior

affinities

average_degree

communities_snr

continuous_overlap

discrete_overlap

effective_snr

evaluate_amp

features_snr

init_amp

nb_features

run_amp

semisupervised_loss_amp

update_amp!

StochasticBlockModelVariants.AbstractLatents — Type

abstract type AbstractLatents

Abstract supertype for SBM communities and latent weights / centroids.

Subtypes

LatentsCSBM

LatentsGLMSBM

StochasticBlockModelVariants.AbstractObservations — Type

abstract type AbstractObservations

Abstract supertype for SBM graph and node feature observations.

Subtypes

ObservationsCSBM

ObservationsGLMSBM

StochasticBlockModelVariants.AbstractSBM — Type

abstract type AbstractSBM

Abstract supertype for Stochastic Block Models with additional node features.

Subtypes

CSBM

GLMSBM

StochasticBlockModelVariants.CSBM — Type

struct CSBM{R} <: AbstractSBM

A generative model for graphs with node features, which combines a Stochastic Block Model with a mixture of Gaussians.

Reference: https://arxiv.org/abs/2306.07948

Fields

N::Int64: graph size

P::Int64: feature dimension

d::Any: average degree

λ::Any: SNR of the communities

μ::Any: SNR of the features

ρ::Any: fraction of node assignments observed

StochasticBlockModelVariants.GLMSBM — Type

struct GLMSBM{R<:Real, D} <: AbstractSBM

A generative model for graphs with node features, which combines a Generalized Linear Model with a Stochastic Block Model.

Reference: https://arxiv.org/abs/2303.09995

Fields

N::Int64: graph size

M::Int64: feature dimension

c::Real: average degree

λ::Real: SNR of the communities

ρ::Real: fraction of nodes revealed

Pʷ::Any: prior on the weights

StochasticBlockModelVariants.LatentsCSBM — Type

struct LatentsCSBM{R<:Real} <: StochasticBlockModelVariants.AbstractLatents

The hidden variables generated by sampling from a CSBM.

Fields

u::Vector{Int64}: community assignments, length N

v::Vector{R} where R<:Real: feature centroids, length P

StochasticBlockModelVariants.LatentsGLMSBM — Type

struct LatentsGLMSBM{R<:Real} <: StochasticBlockModelVariants.AbstractLatents

The hidden variables generated by sampling from a GLMSBM.

Fields

s::Vector{Int64}: community assignments, length N

w::Vector{R} where R<:Real: feature weights, length M

StochasticBlockModelVariants.MarginalsCSBM — Type

struct MarginalsCSBM{R}

Fields

û::Vector: posterior mean of u, length N

v̂::Vector: posterior mean of v, length P

û_no_feat::Vector: posterior mean of u if there were no features, length N (aka Bᵥ)

v̂_no_comm::Vector: posterior mean of v if there were no communities, length P (aka Bᵤ)

h̃₊::Vector: individual external field for u=1, length N

h̃₋::Vector: individual external field for u=-1, length N

χe₊::Dict{Tuple{Int64, Int64}}: messages about the marginal distribution of u, size (N, N)

χ₊::Vector: marginal probability of u=1, length N

StochasticBlockModelVariants.MarginalsGLMSBM — Type

struct MarginalsGLMSBM{R<:Real}

Fields

ŝ::Vector{R} where R<:Real: length N

ŵ::Vector{R} where R<:Real: length M

v::Vector{R} where R<:Real: length M, alias a

Γ::Vector{R} where R<:Real: length M

ω::Vector{R} where R<:Real: length N

gₒ::Vector{R} where R<:Real: length N

h::Vector{R} where R<:Real: length 2

ψl::Matrix{R} where R<:Real: size (2, N)

χe::Dict{Tuple{Int64, Int64, Int64}, R} where R<:Real: size (2, N, N)

χl::Matrix{R} where R<:Real: size (2, N)

χ::Matrix{R} where R<:Real: size (2, N)

StochasticBlockModelVariants.ObservationsCSBM — Type

struct ObservationsCSBM{R<:Real, G<:Graphs.AbstractGraph{Int64}} <: StochasticBlockModelVariants.AbstractObservations

The observations generated by sampling from a CSBM.

Fields

g::Graphs.AbstractGraph{Int64}: undirected unweighted graph with N nodes (~ adjacency matrix A)

Ξ::Vector{Union{Missing, Int64}}: revealed communities ±1 for a fraction ρ of nodes and missing for the rest, length N

B::Matrix{R} where R<:Real: feature matrix, size (P, N)

StochasticBlockModelVariants.ObservationsGLMSBM — Type

struct ObservationsGLMSBM{R<:Real, G<:Graphs.AbstractGraph{Int64}} <: StochasticBlockModelVariants.AbstractObservations

The observations generated by sampling from a GLMSBM.

Fields

g::Graphs.AbstractGraph{Int64}: undirected unweighted graph with N nodes (~ adjacency matrix A)

Ξ::Vector{Union{Missing, Int64}}: revealed communities ±1 for a fraction ρ of nodes and missing for the rest, ngth N

F::Matrix{R} where R<:Real: feature matrix, size (N, M)

Base.length — Function

length(sbm::AbstractSBM)

Return the number of nodes N in the graph.

Base.rand — Function

rand(rng, sbm::AbstractSBM)

Sample observations consisting of a graph and continuous features, return a named tuple (; latents, observations).

StochasticBlockModelVariants.affinities — Method

affinities(sbm::AbstractSBM)

Return a named tuple (; cᵢ, cₒ) containing the affinities inside and outside of a community.

StochasticBlockModelVariants.average_degree — Function

average_degree(sbm::AbstractSBM)

Return the average degre d of a node in the graph.

StochasticBlockModelVariants.communities_snr — Function

communities_snr(sbm::AbstractSBM)

Return the signal-to-noise ratio λ of the communities in the graph.

StochasticBlockModelVariants.continuous_overlap — Method

continuous_overlap(x_true, x_est)

Compute the alignment between two real-valued vectors, up to global sign switch.

StochasticBlockModelVariants.discrete_overlap — Method

discrete_overlap(x_true, x_est)

Compute the alignment between two ±1-valued vectors, up to global sign switch.

StochasticBlockModelVariants.effective_snr — Method

effective_snr(csbm)

Compute the effective SNR λ² + μ² / (N/P).

StochasticBlockModelVariants.evaluate_amp — Method

evaluate_amp(rng, sbm; kwargs...)

Sample observations from sbm and run_amp on them before computing discrete and continuous overlaps with the true latent variables.

Return a tuple (q_dis, q_cont, converged) containing the final overlaps as well as a boolean convergence indicator.

StochasticBlockModelVariants.fraction_observed — Function

fraction_observed(sbm::AbstractSBM)

Return the fraction ρ of community assignments that are observed.

StochasticBlockModelVariants.nb_features — Function

nb_features(sbm::AbstractSBM)

Return the number of features per node, called either P or M depending on the paper.

StochasticBlockModelVariants.run_amp — Method

run_amp(
    rng,
    observations,
    sbm::AbstractSBM;
    init_std,
    max_iterations,
    convergence_threshold,
    recent_past,
    damping,
    show_progress
)

Run AMP-BP for sbm based on a set of observations to estimate the discrete communities and the continuous weights / centroids.

Return a tuple (discrete_history, continuous_history, converged) containing the history of latent variable marginals as well as a boolean convergence indicator.

Keyword arguments

init_std: Noise level used to initialize some messages

max_iterations: Maximum number of iterations allowed

convergence_threshold: Lower bound that the standard deviation of recent latent estimates has to reach for the algorithm to have converged

recent_past: Number of instants taken into account when computing the recent standard deviation

damping: Fraction in [0, 1] telling us how much the previous message is copied into the next message

show_progress: Whether to display a progress bar with convergence statistics

StochasticBlockModelVariants.sample_graph — Method

sample_graph(rng, sbm::AbstractSBM, communities::Vector{<:Integer})

Sample a graph g based on known community assignments.

StochasticBlockModelVariants.sample_mask — Method

sample_mask(rng, sbm::AbstractSBM, communities::Vector{<:Integer})

Sample a vector Ξ (Xi) whose components are equal to the community assignments with probability ρ and equal to missing with probability 1-ρ.