StochasticBlockModelVariants

Documentation for StochasticBlockModelVariants.

StochasticBlockModelVariants.AbstractSBM

StochasticBlockModelVariants.CSBM

StochasticBlockModelVariants.GLMSBM

StochasticBlockModelVariants.LatentsCSBM

StochasticBlockModelVariants.LatentsGLMSBM

StochasticBlockModelVariants.MarginalsCSBM

StochasticBlockModelVariants.ObservationsCSBM

StochasticBlockModelVariants.ObservationsGLMSBM

Base.length

Base.rand

Base.rand

StochasticBlockModelVariants.affinities

StochasticBlockModelVariants.average_degree

StochasticBlockModelVariants.communities_snr

StochasticBlockModelVariants.effective_snr

StochasticBlockModelVariants.fraction_observed

StochasticBlockModelVariants.nb_features

StochasticBlockModelVariants.sample_graph

StochasticBlockModelVariants.sample_mask

StochasticBlockModelVariants.AbstractSBM — Type

AbstractSBM

Abstract supertype for Stochastic Block Models with additional node features.

StochasticBlockModelVariants.CSBM — Type

CSBM

A generative model for graphs with node features, which combines a Stochastic Block Model with a mixture of Gaussians.

Reference: https://arxiv.org/abs/2306.07948

Fields

N: graph size

P: feature dimension

d: average degree

λ: SNR of the communities

μ: SNR of the features

ρ: fraction of node assignments observed

StochasticBlockModelVariants.GLMSBM — Type

GLMSBM

A generative model for graphs with node features, which combines a Generalized Linear Model with a Stochastic Block Model.

Reference: https://arxiv.org/abs/2303.09995

Fields

N: graph size

M: feature dimension

c: average degree

λ: SNR of the communities

ρ: fraction of nodes revealed

Pʷ: prior on the weights

StochasticBlockModelVariants.LatentsCSBM — Type

LatentsCSBM

The hidden variables generated by sampling from a CSBM.

Fields

u::Vector: community assignments, length N

v::Vector: feature centroids, length P

StochasticBlockModelVariants.LatentsGLMSBM — Type

LatentsGLMSBM

The hidden variables generated by sampling from a GLMSBM.

Fields

s::Vector: community assignments, length N

w::Vector: feature weights, length M

StochasticBlockModelVariants.MarginalsCSBM — Type

MarginalsCSBM

Fields

û::Vector: posterior mean of u, length N

v̂::Vector: posterior mean of v, length P

û_no_feat::Vector: posterior mean of u if there were no features, length N (aka Bᵥ)

v̂_no_comm::Vector: posterior mean of v if there were no communities, length P (aka Bᵤ)

h̃₊::Vector: individual external field for u=1, length N

h̃₋::Vector: individual external field for u=-1, length N

χe₊::Dict: messages about the marginal distribution of u, size (N, N)

χ₊::Vector: marginal probability of u=1, length N

StochasticBlockModelVariants.ObservationsCSBM — Type

ObservationsCSBM

The observations generated by sampling from a CSBM.

Fields

g::AbstractGraph: undirected unweighted graph with N nodes (~ adjacency matrix A)

Ξ::Vector: revealed communities ±1 for a fraction ρ of nodes and missing for the rest, length N

B::Matrix: feature matrix, size (P, N)

StochasticBlockModelVariants.ObservationsGLMSBM — Type

ObservationsGLMSBM

The observations generated by sampling from a GLMSBM.

Fields

g::AbstractGraph: undirected unweighted graph with N nodes (~ adjacency matrix A)

Ξ::Vector: revealed communities ±1 for a fraction ρ of nodes and missing for the rest, length N

F::Matrix: feature matrix, size (M, N)

Base.length — Function

length(sbm)

Return the number of nodes N in the graph.

Base.rand — Method

rand(rng, csbm)

Sample from a CSBM and return a named tuple (; latents, observations).

Base.rand — Method

rand(rng, glmsbm)

Sample from a GLMSBM and return a named tuple (; latents, observations).

StochasticBlockModelVariants.affinities — Method

affinities(sbm)

Return a named tuple (; cᵢ, cₒ) containing the affinities inside and outside of a community.

StochasticBlockModelVariants.average_degree — Function

average_degree(sbm)

Return the average degre d of a node in the graph.

StochasticBlockModelVariants.communities_snr — Function

communities_snr(sbm)

Return the signal-to-noise ratio λ of the communities in the graph.

StochasticBlockModelVariants.effective_snr — Method

effective_snr(csbm)

Compute the effective SNR λ² + μ² / (N/P).

StochasticBlockModelVariants.fraction_observed — Function

fraction_observed(sbm)

Return the fraction ρ of community assignments that are observed.

StochasticBlockModelVariants.nb_features — Function

nb_features(sbm)

Return the number of features for each node in the graph.

StochasticBlockModelVariants.sample_graph — Method

sample_graph(rng, sbm, communities)

Sample a graph g from an SBM based on known community assignments.

StochasticBlockModelVariants.sample_mask — Method

sample_mask(rng, sbm, communities)

Sample a vector Ξ (Xi) whose components are equal to the community assignments with probability ρ and equal to missing with probability 1-ρ.