StochasticBlockModelVariants

Documentation for StochasticBlockModelVariants.

StochasticBlockModelVariants.AMPStorage

StochasticBlockModelVariants.AMPTempStorage

StochasticBlockModelVariants.ContextualSBM

StochasticBlockModelVariants.ContextualSBMLatents

StochasticBlockModelVariants.ContextualSBMObservations

Base.rand

StochasticBlockModelVariants.AMPStorage — Type

AMPStorage

Fields

û::Vector: posterior mean of u, length N

v̂::Vector: posterior mean of v, length P

χ₊e::AbstractMatrix: message about the marginal distribution of u, size (N, N)

StochasticBlockModelVariants.AMPTempStorage — Type

AMPTempStorage

Fields

û_no_feat::Vector: posterior mean of u if there were no features, length N (aka Bᵥ)

v̂_no_graph::Vector: posterior mean of v if there were no graph, length P (aka Bᵤ)

h̃₊::Vector: individual external field for u=1, length N

h̃₋::Vector: individual external field for u=-1, length N

χ₊::Vector: marginal probability of u=1, length N

StochasticBlockModelVariants.ContextualSBM — Type

ContextualSBM

Fields

d: average degree

λ: SNR of the graph

μ: SNR of the features

N: graph size

P: feature dimension

StochasticBlockModelVariants.ContextualSBMLatents — Type

ContextualSBMLatents

Fields

u::Vector: community assignments, length N

v::Vector: feature centroids, length P

StochasticBlockModelVariants.ContextualSBMObservations — Type

ContextualSBMObservations

Fields

A::AbstractMatrix: adjacency matrix, size (N, N)

B::Matrix: feature matrix, size (P, N)

G::AbstractGraph: unweighted graph generated from A

Base.rand — Method

rand(rng, csbm)

Simulate a contextual SBM, return a named tuple (; latents, observations).