StochasticBlockModelVariants

Documentation for StochasticBlockModelVariants.

StochasticBlockModelVariants.AMPMarginals

StochasticBlockModelVariants.AMPStorage

StochasticBlockModelVariants.ContextualSBM

StochasticBlockModelVariants.ContextualSBMLatents

StochasticBlockModelVariants.ContextualSBMObservations

Base.rand

StochasticBlockModelVariants.affinities

StochasticBlockModelVariants.effective_snr

StochasticBlockModelVariants.AMPMarginals — Type

AMPMarginals

Fields

û::Vector: posterior mean of u, length N

v̂::Vector: posterior mean of v, length P

χ₊e::Dict: messages about the marginal distribution of u, size (N, N)

StochasticBlockModelVariants.AMPStorage — Type

AMPStorage

Fields

û_no_feat::Vector: posterior mean of u if there were no features, length N (aka Bᵥ)

v̂_no_comm::Vector: posterior mean of v if there were no communities, length P (aka Bᵤ)

h̃₊::Vector: individual external field for u=1, length N

h̃₋::Vector: individual external field for u=-1, length N

χ₊::Vector: marginal probability of u=1, length N

StochasticBlockModelVariants.ContextualSBM — Type

ContextualSBM

A generative model for graphs with node features, which combines the Stochastic Block Model with a mixture of Gaussians.

Reference: https://arxiv.org/abs/2306.07948

Fields

N: graph size

P: feature dimension

d: average degree

λ: SNR of the communities

μ: SNR of the features

ρ: fraction of nodes revealed

StochasticBlockModelVariants.ContextualSBMLatents — Type

ContextualSBMLatents

The hidden variables generated by sampling from a ContextualSBM.

Fields

u::Vector: community assignments, length N

v::Vector: feature centroids, length P

StochasticBlockModelVariants.ContextualSBMObservations — Type

ContextualSBMObservations

The observations generated by sampling from a ContextualSBM.

Fields

A::AbstractMatrix: symmetric boolean adjacency matrix, size (N, N)

g::AbstractGraph: undirected unweighted graph generated from A

B::Matrix: feature matrix, size (P, N)

Ξ::Vector: revealed communities ±1 for a fraction of nodes and 0 for the rest, length N

Base.rand — Method

rand(rng, csbm)

Sample from a ContextualSBM and return a named tuple (; latents, observations).

StochasticBlockModelVariants.affinities — Method

affinities(csbm)

Return a named tuple (; cᵢ, cₒ) containing the affinities inside and outside of a community.

StochasticBlockModelVariants.effective_snr — Method

effective_snr(csbm)

Compute the effective SNR λ² + μ² / (N/P).