StochasticBlockModelVariants

Documentation for StochasticBlockModelVariants.

StochasticBlockModelVariants.StochasticBlockModelVariants

StochasticBlockModelVariants.AbstractSBM

StochasticBlockModelVariants.CSBM

StochasticBlockModelVariants.GLMSBM

StochasticBlockModelVariants.LatentsCSBM

StochasticBlockModelVariants.LatentsGLMSBM

StochasticBlockModelVariants.MarginalsCSBM

StochasticBlockModelVariants.MarginalsGLMSBM

StochasticBlockModelVariants.ObservationsCSBM

StochasticBlockModelVariants.ObservationsGLMSBM

Base.length

Base.rand

Base.rand

StochasticBlockModelVariants.affinities

StochasticBlockModelVariants.average_degree

StochasticBlockModelVariants.communities_snr

StochasticBlockModelVariants.effective_snr

StochasticBlockModelVariants.fraction_observed

StochasticBlockModelVariants.nb_features

StochasticBlockModelVariants.sample_graph

StochasticBlockModelVariants.sample_mask

StochasticBlockModelVariants.StochasticBlockModelVariants — Module

StochasticBlockModelVariants

A package for inference in SBMs with node features using message-passing algorithms.

Exports

CSBM

GLMSBM

LatentsCSBM

LatentsGLMSBM

ObservationsCSBM

ObservationsGLMSBM

affinities

average_degree

communities_snr

effective_snr

evaluate_amp

features_snr

init_amp

nb_features

overlaps

run_amp

update_amp!

StochasticBlockModelVariants.AbstractSBM — Type

abstract type AbstractSBM

Abstract supertype for Stochastic Block Models with additional node features.

StochasticBlockModelVariants.CSBM — Type

struct CSBM{R} <: StochasticBlockModelVariants.AbstractSBM

A generative model for graphs with node features, which combines a Stochastic Block Model with a mixture of Gaussians.

Reference: https://arxiv.org/abs/2306.07948

Fields

N::Int64: graph size

P::Int64: feature dimension

d::Any: average degree

λ::Any: SNR of the communities

μ::Any: SNR of the features

ρ::Any: fraction of node assignments observed

StochasticBlockModelVariants.GLMSBM — Type

struct GLMSBM{R<:Real, D} <: StochasticBlockModelVariants.AbstractSBM

A generative model for graphs with node features, which combines a Generalized Linear Model with a Stochastic Block Model.

Reference: https://arxiv.org/abs/2303.09995

Fields

N::Int64: graph size

M::Int64: feature dimension

c::Real: average degree

λ::Real: SNR of the communities

ρ::Real: fraction of nodes revealed

Pʷ::Any: prior on the weights

StochasticBlockModelVariants.LatentsCSBM — Type

struct LatentsCSBM{R<:Real}

The hidden variables generated by sampling from a CSBM.

Fields

u::Vector{Int64}: community assignments, length N

v::Vector{R} where R<:Real: feature centroids, length P

StochasticBlockModelVariants.LatentsGLMSBM — Type

struct LatentsGLMSBM{R<:Real}

The hidden variables generated by sampling from a GLMSBM.

Fields

s::Vector{Int64}: community assignments, length N

w::Vector{R} where R<:Real: feature weights, length M

StochasticBlockModelVariants.MarginalsCSBM — Type

struct MarginalsCSBM{R}

Fields

û::Vector: posterior mean of u, length N

v̂::Vector: posterior mean of v, length P

û_no_feat::Vector: posterior mean of u if there were no features, length N (aka Bᵥ)

v̂_no_comm::Vector: posterior mean of v if there were no communities, length P (aka Bᵤ)

h̃₊::Vector: individual external field for u=1, length N

h̃₋::Vector: individual external field for u=-1, length N

χe₊::Dict{Tuple{Int64, Int64}}: messages about the marginal distribution of u, size (N, N)

χ₊::Vector: marginal probability of u=1, length N

StochasticBlockModelVariants.MarginalsGLMSBM — Type

struct MarginalsGLMSBM{R<:Real}

Fields

ŝ::Vector{R} where R<:Real: length N

ŵ::Vector{R} where R<:Real: length M

v::Vector{R} where R<:Real: length M, alias a

Γ::Vector{R} where R<:Real: length M

ω::Vector{R} where R<:Real: length N

gₒ::Vector{R} where R<:Real: length N

ψl::Matrix{R} where R<:Real: size (2, N)

χe::Dict{Tuple{Int64, Int64, Int64}, R} where R<:Real: size (2, N, N)

χl::Matrix{R} where R<:Real: size (2, N)

χ::Matrix{R} where R<:Real: size (2, N)

StochasticBlockModelVariants.ObservationsCSBM — Type

struct ObservationsCSBM{R<:Real, G<:Graphs.AbstractGraph{Int64}}

The observations generated by sampling from a CSBM.

Fields

g::Graphs.AbstractGraph{Int64}: undirected unweighted graph with N nodes (~ adjacency matrix A)

Ξ::Vector{Union{Missing, Int64}}: revealed communities ±1 for a fraction ρ of nodes and missing for the rest, length N

B::Matrix{R} where R<:Real: feature matrix, size (P, N)

StochasticBlockModelVariants.ObservationsGLMSBM — Type

struct ObservationsGLMSBM{R<:Real, G<:Graphs.AbstractGraph{Int64}}

The observations generated by sampling from a GLMSBM.

Fields

g::Graphs.AbstractGraph{Int64}: undirected unweighted graph with N nodes (~ adjacency matrix A)

Ξ::Vector{Union{Missing, Int64}}: revealed communities ±1 for a fraction ρ of nodes and missing for the rest, ngth N

F::Matrix{R} where R<:Real: feature matrix, size (M, N)

Base.length — Function

length(sbm::AbstractSBM)

Return the number of nodes N in the graph.

Base.rand — Method

rand(rng, csbm)

Sample from a CSBM and return a named tuple (; latents, observations).

Base.rand — Method

rand(rng, glmsbm)

Sample from a GLMSBM and return a named tuple (; latents, observations).

StochasticBlockModelVariants.affinities — Method

affinities(sbm)

Return a named tuple (; cᵢ, cₒ) containing the affinities inside and outside of a community.

StochasticBlockModelVariants.average_degree — Function

average_degree(sbm::AbstractSBM)

Return the average degre d of a node in the graph.

StochasticBlockModelVariants.communities_snr — Function

communities_snr(sbm::AbstractSBM)

Return the signal-to-noise ratio λ of the communities in the graph.

StochasticBlockModelVariants.effective_snr — Method

effective_snr(csbm)

Compute the effective SNR λ² + μ² / (N/P).

StochasticBlockModelVariants.fraction_observed — Function

fraction_observed(sbm)

Return the fraction ρ of community assignments that are observed.

StochasticBlockModelVariants.nb_features — Function

nb_features(sbm::AbstractSBM)

Return the number of nodes N in the graph.

StochasticBlockModelVariants.sample_graph — Method

sample_graph(rng, sbm, communities)

Sample a graph g from an SBM based on known community assignments.

StochasticBlockModelVariants.sample_mask — Method

sample_mask(rng, sbm, communities)

Sample a vector Ξ (Xi) whose components are equal to the community assignments with probability ρ and equal to missing with probability 1-ρ.